move remaining code from sentry.tasks.integrations.github to sentry.integrations.github.tasks
